@@ -1,7 +1,7 @@
 export const COMMON_CONFIG = {
     apiKey: "AIzaSyBVSy3YpkVGiKXbbxeK0qBnu3-MNZ9UIjA",
     authDomain: "angularfire2-test.firebaseapp.com",
-    databaseURL: "https://angularfire2-test.firebaseio.com",
+    databaseURL: "http://localhost:9000",
     projectId: "angularfire2-test",
     storageBucket: "angularfire2-test.appspot.com",
     messagingSenderId: "920323787688",
